https://youtrack.jetbrains.com/issue/CMP-9400/Merge-Jetpack-Compose-1.11.0-alpha02

| GroupId | ReleaseVersion | ReleaseSHA | ReleaseBuildId | ReleaseDate |
| --- | --- | --- | --- | --- |
| androidx.compose.animation | 1.11.0-alpha02 |
ec985eed3cba8444e5aaa52a748333397a1298f3 | 14574256 | 12/17/2025 |
| androidx.compose.foundation | 1.11.0-alpha02 |
ec985eed3cba8444e5aaa52a748333397a1298f3 | 14574256 | 12/17/2025 |
| androidx.compose.material | 1.11.0-alpha02 |
ec985eed3cba8444e5aaa52a748333397a1298f3 | 14574256 | 12/17/2025 |
| androidx.compose.runtime | 1.11.0-alpha02 |
ec985eed3cba8444e5aaa52a748333397a1298f3 | 14574256 | 12/17/2025 |
| androidx.compose.ui | 1.11.0-alpha02 |
ec985eed3cba8444e5aaa52a748333397a1298f3 | 14574256 | 12/17/2025 |

## Release Notes
N/A

+++ compose/remote/integration-tests/player-view-demos/src/main/java/androidx/compose/remote/integration/view/demos/examples/Demo.kt
@@ -208,7 +208,7 @@ fun plot2(): RemoteComposeWriter {
                                 (minX + maxX) / 2,
                                 min = minX,
                                 max = maxX,
-                                TouchExpression.STOP_INSTANTLY.toInt(),
+                                TouchExpression.STOP_INSTANTLY,
                                 0f,
                                 0,
                                 null,
